requirements.txt
fastapi[all]==0.111.1
uvicorn==0.30.3
pydantic-settings==2.4.0
SQLAlchemy==2.0.31
alembic==1.13.2
asyncpg==0.29.0
black==24.10.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
redis==5.0.8
fastapi-cache2==0.2.2
celery==5.4.0
pillow==10.4.0
ruff
